+++ tests/testdata/other/jobSummary/single_issue.md
@@ -1,3 +1,1 @@
-```
-❌ Found 3 Secrets (2 High, 1 Low)
-```
+<pre>❌ Vulnerabilities found 3<br>└── 3 Secrets 🔴 <span style="color:red">2 High</span><br>              🟡 <span style="color:yellow">1 Low</span></pre>
