+++ utils/resultwriter.go
@@ -4,11 +4,9 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"path/filepath"
 	"strconv"
 	"strings"
 
-	"github.com/jfrog/gofrog/datastructures"
 	"github.com/jfrog/jfrog-cli-core/v2/common/format"
 	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
 	"github.com/jfrog/jfrog-cli-security/formats"
@@ -18,7 +16,6 @@ import (
 	"github.com/jfrog/jfrog-client-go/utils/log"
 	"github.com/jfrog/jfrog-client-go/xray/services"
 	"github.com/owenrumney/go-sarif/v2/sarif"
-	"golang.org/x/exp/maps"
 	"golang.org/x/exp/slices"
 )
 
@@ -526,296 +523,6 @@ func writeJsonResults(results *Results) (resultsPath string, err error) {
 	}
 	resultsPath = out.Name()
 	return
-}
-
-func GetSummaryString(summaries ...formats.SummaryResults) (str string, err error) {
-	parsed := 0
-	singleScan := isSingleCommandAndScan(summaries...)
-	wd, err := coreutils.GetWorkingDirectory()
-	if err != nil {
-		return
-	}
-	for i := range summaries {
-		if !singleScan {
-			updateSummaryNamesToRelativePath(&summaries[i], wd)
-		}
-		for _, scan := range summaries[i].Scans {
-			if parsed > 0 {
-				str += "\n"
-			}
-			str += GetScanSummaryString(scan, singleScan)
-			parsed++
-		}
-	}
-	return
-}
-
-func isSingleCommandAndScan(summaries ...formats.SummaryResults) bool {
-	if len(summaries) != 1 {
-		return false
-	}
-	if len(summaries[0].Scans) != 1 {
-		return false
-	}
-	// One command and one scan
-	return true
-}
-
-func GetScanSummaryString(summary formats.ScanSummaryResult, singleData bool) (content string) {
-	if !summary.HasIssues() {
-		if singleData {
-			return "✅ No vulnerabilities were found"
-		}
-		return fmt.Sprintf("✅ %s", summary.Target)
-	}
-	// Handle display of issues
-	content = "❌"
-	if !singleData {
-		content += fmt.Sprintf(" %s:", summary.Target)
-	}
-	content += " Found"
-	subScansWithIssues := summary.GetSubScansWithIssues()
-	if len(subScansWithIssues) == 1 {
-		content += getSubScanSummaryString(summary, subScansWithIssues[0])
-		return
-	}
-	// Multiple sub scans with issues
-	content += fmt.Sprintf(" %d vulnerabilities\n", summary.GetTotalIssueCount())
-	content += getSubScanSummaryString(summary, subScansWithIssues...)
-	return
-}
-
-func getSummaryListItemPrefix(index, total, subListLevel int) (content string) {
-	for i := 0; i < subListLevel; i++ {
-		content += "    "
-	}
-	if subListLevel == 0 && total < 2 {
-		return " "
-	}
-	if index == total-1 {
-		content += "└── "
-		return
-	}
-	content += "├── "
-	return
-}
-
-func getSubScanSummaryCountsString(summary formats.ScanSummaryResult, subScanType formats.SummarySubScanType) (content string) {
-	switch subScanType {
-	case formats.ScaScan:
-		content += GetScaSummaryCountString(*summary.ScaScanResults)
-	case formats.IacScan:
-		content += GetSummaryCountString(*summary.IacScanResults)
-	case formats.SecretsScan:
-		content += GetSummaryCountString(*summary.SecretsScanResults)
-	case formats.SastScan:
-		content += GetSummaryCountString(*summary.SastScanResults)
-	}
-	return
-}
-
-func hasApplicableDataToDisplayInSummary(summary formats.ScaSummaryCount) bool {
-	for _, statuses := range summary {
-		sorted := getSummarySortedKeysToDisplay(maps.Keys(statuses)...)
-		for _, status := range sorted {
-			if _, ok := statuses[status]; ok && statuses[status] > 0 {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func GetScaSummaryCountString(summary formats.ScaSummaryCount) (content string) {
-	severityCount := len(summary)
-	if severityCount == 0 {
-		return
-	}
-	if !hasApplicableDataToDisplayInSummary(summary) {
-		return GetSummaryCountString(summary.GetSeverityCountsWithoutStatus())
-	}
-	// Display contextual-analysis details
-	keys := getSummarySortedKeysToDisplay(maps.Keys(summary)...)
-	for i, severity := range keys {
-		statusCounts := summary[severity]
-		content += fmt.Sprintf("\n%s%d %s%s",
-			getSummaryListItemPrefix(i, severityCount, 1),
-			statusCounts.GetTotal(),
-			severity,
-			GetSummaryCountString(statusCounts),
-		)
-	}
-	return
-}
-
-func getSummarySortedKeysToDisplay(keys ...string) (sorted []string) {
-	if len(keys) == 0 {
-		return
-	}
-	keysSet := datastructures.MakeSetFromElements(keys...)
-	allowedSorted := []string{
-		"Critical", "High", "Medium", "Low", "Unknown",
-		string(Applicable), string(NotApplicable),
-	}
-	for _, key := range allowedSorted {
-		if keysSet.Exists(key) {
-			sorted = append(sorted, key)
-		}
-	}
-	return
-}
-
-func GetSummaryCountString(summary formats.SummaryCount) string {
-	if len(summary) == 0 {
-		return ""
-	}
-	// sort and filter
-	keys := getSummarySortedKeysToDisplay(maps.Keys(summary)...)
-	if len(keys) == 0 {
-		return ""
-	}
-	content := ""
-	for i, key := range keys {
-		if i > 0 {
-			content += ", "
-		}
-		content += fmt.Sprintf("%d %s", summary[key], key)
-	}
-	return fmt.Sprintf(" (%s)", content)
-}
-
-func updateSummaryNamesToRelativePath(summary *formats.SummaryResults, wd string) {
-	for i, scan := range summary.Scans {
-		if scan.Target == "" {
-			continue
-		}
-		if !strings.HasPrefix(scan.Target, wd) {
-			continue
-		}
-		if scan.Target == wd {
-			summary.Scans[i].Target = filepath.Base(wd)
-		}
-		summary.Scans[i].Target = strings.TrimPrefix(scan.Target, wd)
-	}
-}
-
-func getSubScanSummaryString(summary formats.ScanSummaryResult, subScanTypes ...formats.SummarySubScanType) (content string) {
-	totalSubScans := len(subScanTypes)
-	for i, subScanType := range subScanTypes {
-		// Prefix
-		if i > 0 {
-			content += "\n"
-		}
-		content += fmt.Sprintf("%s%d ", getSummaryListItemPrefix(i, totalSubScans, 0), summary.GetSubScanTotalIssueCount(subScanType))
-		switch subScanType {
-		case formats.ScaScan:
-			content += "SCA vulnerabilities"
-		case formats.IacScan:
-			content += "IAC vulnerabilities"
-		case formats.SecretsScan:
-			content += "Secrets"
-		case formats.SastScan:
-			content += "SAST vulnerabilities"
-		}
-		content += getSubScanSummaryCountsString(summary, subScanType)
-	}
-	return
-}
-
-func getScanSummary(extendedScanResults *ExtendedScanResults, scaResults ...ScaScanResult) (summary formats.ScanSummaryResult) {
-	if len(scaResults) == 1 {
-		summary.Target = scaResults[0].Target
-	}
-	if extendedScanResults == nil {
-		summary.ScaScanResults = getScaSummaryResults(&scaResults)
-		return
-	}
-	summary.ScaScanResults = getScaSummaryResults(&scaResults, extendedScanResults.ApplicabilityScanResults...)
-	summary.IacScanResults = getJASSummaryCount(extendedScanResults.IacScanResults...)
-	summary.SecretsScanResults = getJASSummaryCount(extendedScanResults.SecretsScanResults...)
-	summary.SastScanResults = getJASSummaryCount(extendedScanResults.SastScanResults...)
-	return
-}
-
-type SeverityWithApplicable struct {
-	SeverityInfo        *TableSeverity
-	ApplicabilityStatus ApplicabilityStatus
-}
-
-func getCveId(cve services.Cve, defaultIssueId string) string {
-	if cve.Id == "" {
-		return defaultIssueId
-	}
-	return cve.Id
-}
-
-func getUniqueVulnerabilitiesInfo(cves []services.Cve, issueId, severity string, components map[string]services.Component, applicableRuns ...*sarif.Run) (uniqueFindings map[string]SeverityWithApplicable) {
-	uniqueFindings = map[string]SeverityWithApplicable{}
-	for _, cve := range cves {
-		cveId := getCveId(cve, issueId)
-		for compId := range components {
-			applicableStatus := NotScanned
-			if applicableInfo := getCveApplicabilityField(cveId, applicableRuns, components); applicableInfo != nil {
-				applicableStatus = convertToApplicabilityStatus(applicableInfo.Status)
-			}
-			uniqueFindings[cveId+compId] = SeverityWithApplicable{SeverityInfo: GetSeverity(severity, applicableStatus), ApplicabilityStatus: applicableStatus}
-		}
-	}
-	return
-}
-
-func getScaSummaryResults(scaScanResults *[]ScaScanResult, applicableRuns ...*sarif.Run) *formats.ScaSummaryCount {
-	uniqueFindings := map[string]SeverityWithApplicable{}
-	if len(*scaScanResults) == 0 {
-		return nil
-	}
-	// Aggregate unique findings
-	for _, scaResult := range *scaScanResults {
-		for _, xrayResult := range scaResult.XrayResults {
-			for _, vulnerability := range xrayResult.Vulnerabilities {
-				vulUniqueFindings := getUniqueVulnerabilitiesInfo(vulnerability.Cves, vulnerability.IssueId, vulnerability.Severity, vulnerability.Components, applicableRuns...)
-				for key, value := range vulUniqueFindings {
-					uniqueFindings[key] = value
-				}
-			}
-			for _, violation := range xrayResult.Violations {
-				vioUniqueFindings := getUniqueVulnerabilitiesInfo(violation.Cves, violation.IssueId, violation.Severity, violation.Components, applicableRuns...)
-				for key, value := range vioUniqueFindings {
-					uniqueFindings[key] = value
-				}
-			}
-		}
-	}
-	// Create summary
-	summary := formats.ScaSummaryCount{}
-	for _, severityWithApplicable := range uniqueFindings {
-		severity := severityWithApplicable.SeverityInfo.Severity
-		status := severityWithApplicable.ApplicabilityStatus.String()
-		if _, ok := summary[severity]; !ok {
-			summary[severity] = formats.SummaryCount{}
-		}
-		summary[severity][status]++
-	}
-	return &summary
-}
-
-func getJASSummaryCount(runs ...*sarif.Run) *formats.SummaryCount {
-	if len(runs) == 0 {
-		return nil
-	}
-	count := formats.SummaryCount{}
-	issueToSeverity := map[string]string{}
-	for _, run := range runs {
-		for _, result := range run.Results {
-			for _, location := range result.Locations {
-				issueToSeverity[GetLocationId(location)] = GetResultSeverity(result)
-			}
-		}
-	}
-	for _, severity := range issueToSeverity {
-		count[severity]++
-	}
-	return &count
 }
 
 func PrintJson(output interface{}) error {
